requirements.txt
beautifulsoup4>=4.12.0
feedparser>=6.0.0
matplotlib>=3.8.0
playwright>=1.52.0
PyYAML>=6.0.0
requests>=2.31.0
